separated files in auth packages removed garbage
